Move ensure_pointer() call to oswitch (#2035)
When a view is opened on another output than the focused one, i.e.
with window-rules plugin, focus_output() was calling ensure_pointer()
which was moving the pointer to the center of the output where the
view was opened. Here we move the call to oswitch so this only
happens when the user would want it to.

Fixes #2013.