optimize the GPU usage of workspace-wall (#2132)
In Wayfire 0.7.2 we were using auxilliary buffers to composite the
workspaces before finally drawing them on the screen.

In Wayfire 0.8.0 the behavior changed: all windows were directly
composited on the screen. This introduced highly improved performance
for cases where the workspace contents were changing, because we could
render them at scale. However, it introduced problems with static
workspaces containing multiple windows, because we'd have to composite
them multiple times on each frame.

The new implementation takes a best-of-both-worlds approach. We
composite workspaces to auxilliary buffers, ensuring that we do not
re-composite static surfaces together on each frame.

To ensure that dynamic content also works well, we scale the buffers as
well, if enough of the content has changed so that a full redraw with
a different scale is less expensive than updating the current buffers.
We also have to be careful to avoid visual artifacts (popping etc) when
transitioning between different scales.

Fixes #1940