vswitch: remove "ghost window" on target workspace. (#2106)
When animating, the plugin disables the original view, then adds
a new output node which is rendered as an overlay while the
animation runs. Unfortunately the workspace wall also picks up
the new output node and draws it on the target workspace,
resulting in a duplicate window.

This commit is a simple fix that uses a translation node instead
of an output node. This requires fixing what looks like a missing
return statement in vswitch_overlay_node_t::get_bounding_box.

Fixes #2041