conf: fix reloading on IN_CREATE (#2215)
If the config file is a symlink, wayfire attempts to watch the parent
directory so that, if replaced via a subsequent `ln -sf`, it'll still
be notified. When notified, however, wayfire doesn't immediately
reload the config file. Rather, as this event may be triggered by any
file creation in the directory, we first make sure that the config file
really is the one being created anew.

Unfortunately, ln -sf is special: It'll create a randomly named file in
the directory before renameat(2)ing it to the proper location. So it
does generate an IN_CREATE event, but to know that the file was changed,
we need to additionally monitor for IN_MOVED_TO.

Fixes #2210.