Implement the session lock protocol (#2162)
* Make wlr_{surface,subsurface}_controller_t common utilities.

These classes are useful and needed by any code that deals with
surfaces, because such code usually needs to deal with
subsurfaces as well.

Move the include files into api/wayfire/unstable so they can be
used by code all over the tree, and rename them to reflect the
utility classes that they contain.

* Add a public API to inhibit/uninhibit outputs.

This will allow lockscreens to be implemented in plugins.

* Factor the text label rendering out of wset into a utility class.

This adds a simple node that can display text.

* Add a LOCK scenegraph layer.

This will be used by the upcoming implementation of the
ext-session-lock protocol.

* Implement the session lock protocol

Currently implemented:

- Locking and unlocking.
- Display above all desktop layers, including OVERLAY (but below
  DWIDGET).
- Waiting for the client to display locked surfaces before
  declaring the screen to be locked. If the client does not
  display all surfaces, locks anyway after a timeout.
- Multiple outputs, including outputs being added/removed while
  locking or locked.
- Leaving the screen locked if the client crashes (but allows a
  new client to come in and lock the screen).

TODO:
- Deal with outputs being resized.
- Properly display text if the screenlocker crashes. Currently it
  just displays two a big "explosion" emoji.
- Use a view instead of a plain wlr_surface_node?

Tested with swaylock.

Fixes #1494
Fixes #1358