PR 05 Move safe-list.hpp to (updated) wf-config. (#2578)
* Remove safe-list.hpp from wayfire.

It is now part of wf-config.

* Missing header.

No longer sneaked in through safe-list.hpp.

* Update to latest wf-config subproject.

* Add include directory of wf-config to tests in order to find safe-list.hpp.

* Revert "Add include directory of wf-config to tests in order to find safe-list.hpp."

This reverts commit 005117ab84fdfc25ad18793b3896ee5666bbf893.

Apparently not the way to do it according to soreau.

* The other way.

Just adding include directories would be sufficient, because all we need
is safe-list.hpp, which is header only. This MIGHT drag in shared
libraries and compiler flags and all what not... but maybe it isn't.
Besides, it is only for a test anyway.